src/third_party/blink/web_tests/external/wpt/media-capabilities/idlharness.any.worker-expected.txt

This is a testharness.js-based test.
PASS idlharness
PASS idl_test setup
PASS Partial interface Navigator: original interface defined
PASS Partial interface Navigator: valid exposure set
PASS Partial interface WorkerNavigator: original interface defined
PASS Partial interface WorkerNavigator: valid exposure set
PASS Partial interface Screen: original interface defined
FAIL MediaCapabilities interface: existence and properties of interface object assert_own_property: self does not have own property "MediaCapabilities" expected property "MediaCapabilities" missing
FAIL MediaCapabilities interface object length assert_own_property: self does not have own property "MediaCapabilities" expected property "MediaCapabilities" missing
FAIL MediaCapabilities interface object name assert_own_property: self does not have own property "MediaCapabilities" expected property "MediaCapabilities" missing
FAIL MediaCapabilities interface: existence and properties of interface prototype object assert_own_property: self does not have own property "MediaCapabilities" expected property "MediaCapabilities" missing
FAIL MediaCapabilities interface: existence and properties of interface prototype object's "constructor" property assert_own_property: self does not have own property "MediaCapabilities" expected property "MediaCapabilities" missing
FAIL MediaCapabilities interface: existence and properties of interface prototype object's @@unscopables property assert_own_property: self does not have own property "MediaCapabilities" expected property "MediaCapabilities" missing
FAIL MediaCapabilities interface: operation decodingInfo(MediaDecodingConfiguration) assert_own_property: self does not have own property "MediaCapabilities" expected property "MediaCapabilities" missing
FAIL MediaCapabilities interface: operation encodingInfo(MediaEncodingConfiguration) assert_own_property: self does not have own property "MediaCapabilities" expected property "MediaCapabilities" missing
FAIL MediaCapabilities must be primary interface of navigator.mediaCapabilities assert_equals: wrong typeof object expected "object" but got "undefined"
FAIL Stringification of navigator.mediaCapabilities assert_equals: wrong typeof object expected "object" but got "undefined"
FAIL MediaCapabilities interface: navigator.mediaCapabilities must inherit property "decodingInfo(MediaDecodingConfiguration)" with the proper type assert_equals: wrong typeof object expected "object" but got "undefined"
FAIL MediaCapabilities interface: calling decodingInfo(MediaDecodingConfiguration) on navigator.mediaCapabilities with too few arguments must throw TypeError assert_equals: wrong typeof object expected "object" but got "undefined"
FAIL MediaCapabilities interface: navigator.mediaCapabilities must inherit property "encodingInfo(MediaEncodingConfiguration)" with the proper type assert_equals: wrong typeof object expected "object" but got "undefined"
FAIL MediaCapabilities interface: calling encodingInfo(MediaEncodingConfiguration) on navigator.mediaCapabilities with too few arguments must throw TypeError assert_equals: wrong typeof object expected "object" but got "undefined"
PASS ScreenLuminance interface: existence and properties of interface object
FAIL ScreenLuminance must be primary interface of screen.luminance assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: screen is not defined"
FAIL Stringification of screen.luminance assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: screen is not defined"
FAIL ScreenLuminance interface: screen.luminance must not have property "min" assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: screen is not defined"
FAIL ScreenLuminance interface: screen.luminance must not have property "max" assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: screen is not defined"
FAIL ScreenLuminance interface: screen.luminance must not have property "maxAverage" assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: screen is not defined"
PASS Navigator interface: existence and properties of interface object
FAIL WorkerNavigator interface: attribute mediaCapabilities assert_true: The prototype object must have a property "mediaCapabilities" expected true got false
FAIL WorkerNavigator interface: navigator must inherit property "mediaCapabilities" with the proper type assert_inherits: property "mediaCapabilities" not found in prototype chain
PASS Screen interface: existence and properties of interface object
FAIL Screen interface: screen must not have property "colorGamut" assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: screen is not defined"
FAIL Screen interface: screen must not have property "luminance" assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: screen is not defined"
FAIL Screen interface: screen must not have property "onchange" assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: screen is not defined"
Harness: the test ran to completion.
